some changes

+++ bltin_export.c
@@ -70,13 +70,36 @@ static void	f2(char **argv, char *line)
 	(void)argv;
 }
 
+static void	f3(char **argv, char *line)
+{
+	char	*p;
+
+	if (!line)
+		return ;
+	p = ft_strchr(line, '=');
+	ft_putstr_fd("declare -x ", 1);
+	write(1, line, (p - line) + 1);
+	ft_putchar_fd('\"', 1);
+	if (ft_strlen(p + 1))
+		write(1, p + 1, ft_strlen(p + 1) - 1);
+	ft_putstr_fd("\"\n", 1);
+	(void)argv;
+}
+
 void	minishell_export(char **argv)
 {
 	char	*trc_path;
 
-	env_loop(RC_PATH, argv, f1);
-	env_loop(TRC_PATH, argv, f2);
-	trc_path = ft_strjoin(getenv("HOME"), TRC_PATH);
-	unlink(trc_path);
-	free(trc_path);
+	if (argv[1])
+	{
+		env_loop(RC_PATH, argv, f1);
+		env_loop(TRC_PATH, argv, f2);
+		trc_path = ft_strjoin(getenv("HOME"), TRC_PATH);
+		unlink(trc_path);
+		free(trc_path);
+	}
+	else
+	{
+		env_loop(RC_PATH, argv, f3);
+	}
 }
